# identity/cache: only call SortedList for release