@bimmlerd 2023-08-30
```
goos: linux
goarch: arm64
pkg: github.com/cilium/cilium/pkg/labels
                     │     old     │                 opt                 │
                     │   sec/op    │   sec/op     vs base                │
Labels_SortedList-10   3.279µ ± 6%   2.209µ ± 6%  -32.65% (p=0.000 n=10)

                     │    old     │                opt                 │
                     │    B/op    │    B/op     vs base                │
Labels_SortedList-10   504.0 ± 0%   360.0 ± 0%  -28.57% (p=0.000 n=10)

                     │     old     │                opt                 │
                     │  allocs/op  │ allocs/op   vs base                │
Labels_SortedList-10   13.000 ± 0%   3.000 ± 0%  -76.92% (p=0.000 n=10)
```
for the existing benchmark - I'll write another one which focusses on a the relevant CIDR identity case.