@bimmlerd 2023-09-05
I've pushed the above-suggested change to pull the call to `SortedList` out of the critical section of `lookupOrCreate`. Please take another look.